requirements.txt
Django>=5.0,<6.0
gunicorn>=21.0
requests>=2.31
django-encrypted-model-fields>=0.6
django-apscheduler>=0.6
APScheduler>=3.10,<4.0
python-dotenv>=1.0
whitenoise>=6.6
